OkFiles

Documentation for OkFiles.

OkFiles.datalist

OkFiles.filelist

OkFiles.filelist

OkFiles.filelist

OkFiles.filelistall

OkFiles.filelistall

OkFiles.folderlist

OkFiles.folderlist

OkFiles.folderlist

OkFiles.folderlistall

OkFiles.folderlistall

OkFiles.mv2dir

OkFiles.pathnorepeat

OkFiles.datalist — Method

Example:

 find_in = "D:\GoogleDrive\1Programming\DATA"
 regexppattern = r".*\.mat"
 find = "both"
 method = "readdir"
 O = datalist(regexppattern, find_in, method, find);

Output:

 O.fullpath is a string vector of the paths of the desired files or folders.

OkFiles.filelist — Method

filelist(dir; join=true, sort=true) return the list of files. For its keyword arguments, see readdir.

OkFiles.filelist — Method

filelist(expr::Regex, dir; join=true, sort=true) returns a vector of paths where the file matches expr.

OkFiles.filelist — Method

filelist(; join=true, sort=true) returns the paths for all the files in the current directory.

OkFiles.filelistall — Method

Similar to filelist, filelistall(dir) returns all files in the directory (as well as its subdirectory) dir.

OkFiles.filelistall — Method

Similar to filelist, filelistall(expr::Regex, dir) returns files whose name matches expr in the directory (as well as its subdirectory) dir.

OkFiles.folderlist — Method

folderlist(dir; join=true) return the list of folder but no subfolder under dir. This function uses walkdir.

OkFiles.folderlist — Method

folderlist(expr::Regex, dir; join=true) returns a vector of paths who match expr.

OkFiles.folderlist — Method

folderlist(;kwargs...) returns results in the current directory (pwd)

OkFiles.folderlistall — Method

folderlistall(dir) return the list of folders and all subfolders under dir. For its keyword arguments, see walkdir.

OkFiles.folderlistall — Method

folderlistall() returns results in the current directory (pwd)

OkFiles.mv2dir — Method

mv2dir(item, newdir::AbstractString) moves item to the newdir directory, preseving its original name.

OkFiles.pathnorepeat — Method

pathnorepeat(filepath; suffix_fun = serial_number_4d) returns non-repeated file path, by increasing the serial_number_4d by default. suffix_fun can be assigned as an arbitrary function that suffix_fun(n) returns a string updated by integer n.